codelens-ai/backend/requirements.txt
fastapi>=0.104.0
uvicorn>=0.24.0
python-dotenv>=1.0.0
pydantic>=2.5.0
tree-sitter>=0.21.0
tree-sitter-python>=0.21.0
tree-sitter-javascript>=0.21.0
tree-sitter-typescript>=0.21.0
google-genai>=1.0.0
rank-bm25>=0.2.2
aiosqlite>=0.19.0
redis>=5.0.0
rq>=1.16.0
